Fix a few other warnings

+++ src/shared/Util.hpp
@@ -225,10 +225,10 @@ struct SmallTimeTracker
 
     public:
 
-        SmallTimeTracker(uint32_t expired = 0) : mExpireTime(expired) {}
+        SmallTimeTracker(uint32_t expired = 0) : mExpireTime(static_cast<int32_t>(expired)) {}
 
         void updateTimer(uint32_t diffTime) { mExpireTime -= diffTime; }
-        void resetInterval(uint32_t intervalTime) { mExpireTime = intervalTime; }
+        void resetInterval(uint32_t intervalTime) { mExpireTime = static_cast<int32_t>(intervalTime); }
 
         int32_t getExpireTime() const { return mExpireTime; }
         bool isTimePassed() const { return mExpireTime <= 0; }
